backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
filelock==3.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
fonttools==4.60.1
fsspec==2025.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
PyJWT==2.10.1
pymongo==4.15.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
